Update patch set 1
Patch Set 1:

I think the best option is to make no assumptions about the relation between the user id and group id.

'Group' is a fingerprint hal semantic that describes a set of fingerprints. 

It's the caller's responsibility (in this case the framework) to pass which active set (group) should be used for a given operation.

In the current implementation the framework uses the user id as the group id. So there's a 1 to 1 mapping between user ids and group ids and they happen to be the same but I wouldn't assume it will be the case forever.

Patch-set: 1