Update patch set 1
Patch Set 1:

This makes sense as a last resort but I think we should have an upgrade path in the framework.

In cm-12.1, only the owner had fingerprints so we could trivially convert the data model to the owner gid and add its set of fingerprints to that gid. The framework would call hal->enumerate and fill in framework metadata.

Roman, what do you think?

Patch-set: 1